re fork

+++ src/test/feature/lib/sql_util.cpp
@@ -20,7 +20,7 @@ using std::string;
 namespace hawq {
 namespace test {
 
-SQLUtility::SQLUtility()
+SQLUtility::SQLUtility(SQLUtilityMode mode)
     : testRootPath(getTestRootPath()),
       test_info(::testing::UnitTest::GetInstance()->current_test_info()) {
   auto getConnection = [&] () {
@@ -35,15 +35,36 @@ SQLUtility::SQLUtility()
   };
   getConnection();
 
-  schemaName =
-      string(test_info->test_case_name()) + "_" + test_info->name();
-  exec("DROP SCHEMA IF EXISTS " + schemaName + " CASCADE");
-  exec("CREATE SCHEMA " + schemaName);
+  if (mode == MODE_SCHEMA) {
+    schemaName = string(test_info->test_case_name()) + "_" + test_info->name();
+    databaseName = HAWQ_DB;
+    exec("DROP SCHEMA IF EXISTS " + schemaName + " CASCADE");
+    exec("CREATE SCHEMA " + schemaName);
+    sql_util_mode = MODE_SCHEMA;
+  } else {
+    schemaName = HAWQ_DEFAULT_SCHEMA;
+    databaseName = "db_" + string(test_info->test_case_name()) + "_" + test_info->name();
+    std::transform(databaseName.begin(), databaseName.end(), databaseName.begin(), ::tolower);
+    exec("DROP DATABASE IF EXISTS " + databaseName);
+    exec("CREATE DATABASE " + databaseName);
+    sql_util_mode = MODE_DATABASE;
+  }
 }
 
 SQLUtility::~SQLUtility() {
-  if (!test_info->result()->Failed())
-    exec("DROP SCHEMA " + schemaName + " CASCADE");
+  if (!test_info->result()->Failed()) {
+    if (schemaName != HAWQ_DEFAULT_SCHEMA) {
+      exec("DROP SCHEMA " + schemaName + " CASCADE");
+    }
+
+    if (sql_util_mode ==  MODE_DATABASE) {
+      exec("DROP DATABASE " + databaseName);
+    }
+  }
+}
+
+std::string SQLUtility::getDbName() {
+    return databaseName;
 }
 
 void SQLUtility::exec(const string &sql) {
@@ -58,7 +79,6 @@ string SQLUtility::execute(const string &sql, bool check) {
     EXPECT_EQ(0, conn->getLastStatus()) << conn->getLastResult();
     return "";
   }
-  EXPECT_NE(0,  conn->getLastStatus());
   return conn.get()->getLastResult();
 }
 
@@ -143,9 +163,30 @@ void SQLUtility::execSQLFile(const string &sqlFile,
   }
 }
 
+bool SQLUtility::execSQLFile(const string &sqlFile) {
+  // do precheck for sqlFile
+  if (hawq::test::startsWith(sqlFile, "/"))
+    return false;
+
+  // double check to avoid empty fileBaseName
+  FilePath fp = splitFilePath(sqlFile);
+  if (fp.fileBaseName.empty())
+    return false;
+
+  // outFile is located in the same folder with ansFile
+  string outFileAbsPath = "/tmp/" + fp.fileBaseName + ".out";
+
+  // generate new sql file with set search_path added at the begining
+  const string newSqlFile = generateSQLFile(sqlFile);
+
+  // run sql file and store its result in output file
+  conn->setOutputFile(outFileAbsPath);
+  return conn->runSQLFile(newSqlFile).getLastStatus() == 0 ? true : false;
+}
+
 const string SQLUtility::generateSQLFile(const string &sqlFile) {
   const string originSqlFile = testRootPath + "/" + sqlFile;
-  const string newSqlFile = "/tmp/" + schemaName + ".sql";
+  const string newSqlFile = "/tmp/" + string(test_info->test_case_name()) + "_" + test_info->name() + ".sql";
   std::fstream in;
   in.open(originSqlFile, std::ios::in);
   if (!in.is_open()) {
@@ -156,9 +197,12 @@ const string SQLUtility::generateSQLFile(const string &sqlFile) {
   if (!out.is_open()) {
     EXPECT_TRUE(false) << "Error opening file " << newSqlFile;
   }
-  out << "-- start_ignore" << std::endl
-      << "SET SEARCH_PATH=" + schemaName + ";" << std::endl
-      << "-- end_ignore" << std::endl;
+  out << "-- start_ignore" << std::endl;
+  out << "SET SEARCH_PATH=" + schemaName + ";" << std::endl;
+  if (sql_util_mode ==  MODE_DATABASE) {
+    out << "\\c " << databaseName << std::endl;
+  }
+  out << "-- end_ignore" << std::endl;
   string line;
   while (getline(in, line)) {
     out << line << std::endl;
@@ -210,6 +254,18 @@ std::string SQLUtility::getGUCValue(const std::string &guc) {
   EXPECT_EQ(result.rowCount(), 1);
   std::vector<std::string> row = result.getRows()[0];
   return row[0];
+}
+
+std::string SQLUtility::getQueryResult(const std::string &query) {
+  const hawq::test::PSQLQueryResult &result = executeQuery(query);
+  EXPECT_LE(result.rowCount(), 1);
+  std::string value;
+  if (result.rowCount() == 1)
+  {
+    value = result.getRows()[0][0];
+  }
+
+  return value;
 }
 
 FilePath SQLUtility::splitFilePath(const string &filePath) const {
